+++ packages/studio/CHANGELOG.md
@@ -1,3 +1,5 @@
+## @lunora/studio [1.0.0-alpha.34](https://github.com/anolilab/lunora/compare/%40lunora%2Fstudio%401.0.0-alpha.33...%40lunora%2Fstudio%401.0.0-alpha.34) (2026-07-03)
+
 ## @lunora/studio [1.0.0-alpha.33](https://github.com/anolilab/lunora/compare/%40lunora%2Fstudio%401.0.0-alpha.32...%40lunora%2Fstudio%401.0.0-alpha.33) (2026-07-03)