+++ packages/vite/CHANGELOG.md
@@ -1,3 +1,11 @@
+## @lunora/vite [1.0.0-alpha.49](https://github.com/anolilab/lunora/compare/%40lunora%2Fvite%401.0.0-alpha.48...%40lunora%2Fvite%401.0.0-alpha.49) (2026-07-03)
+
+
+### Dependencies
+
+* **@lunora/config:** upgraded to 1.0.0-alpha.49
+* **@lunora/studio:** upgraded to 1.0.0-alpha.34
+
 ## @lunora/vite [1.0.0-alpha.48](https://github.com/anolilab/lunora/compare/%40lunora%2Fvite%401.0.0-alpha.47...%40lunora%2Fvite%401.0.0-alpha.48) (2026-07-03)